+++ wit/testdata_test.go
@@ -2,8 +2,10 @@ package wit
 
 import (
 	"bytes"
+	"context"
 	"flag"
 	"fmt"
+	"io/fs"
 	"os"
 	"os/exec"
 	"strings"
@@ -11,6 +13,7 @@ import (
 	"testing"
 
 	"go.bytecodealliance.org/internal/relpath"
+	"go.bytecodealliance.org/internal/wasmtools"
 
 	"github.com/sergi/go-diff/diffmatchpatch"
 )
@@ -78,26 +81,33 @@ func TestGoldenWITRoundTrip(t *testing.T) {
 		// t.Skip is not available in TinyGo, requires runtime.Goexit()
 		return
 	}
-	if !canWasmTools() {
-		t.Log("skipping test: wasm-tools not installed or cannot fork/exec (TinyGo)")
+	ctx := context.Background()
+	wasmTools, err := wasmtools.New(ctx)
+	if err != nil {
+		t.Logf("wasm-tools not available: %v", err)
 		return
 	}
-	err := loadTestdata(func(path string, res *Resolve) error {
+
+	err = loadTestdata(func(path string, res *Resolve) error {
 		data := res.WIT(nil, "")
 		t.Run(path, func(t *testing.T) {
-			// Run the generated WIT through wasm-tools to generate JSON.
-			cmd := exec.Command("wasm-tools", "component", "wit", "-j", "--all-features")
-			cmd.Stdin = strings.NewReader(data)
-			stdout := &bytes.Buffer{}
-			stderr := &bytes.Buffer{}
-			cmd.Stdout = stdout
-			cmd.Stderr = stderr
-			err := cmd.Run()
+			args := []string{"component", "wit", "-j", "--all-features"}
+			stdin := strings.NewReader(data)
+
+			fsMap := make(map[fs.FS]string)
+			stdout, stderr, err := wasmTools.Run(ctx, args, stdin, fsMap, &path)
 			if err != nil {
-				t.Error(err, stderr.String())
+				t.Errorf("wasm-tools: %v", err)
 				return
 			}
-
+			if stderr != nil {
+				buf := new(bytes.Buffer)
+				buf.ReadFrom(stderr)
+				if buf.Len() > 0 {
+					t.Error(buf.String())
+					return
+				}
+			}
 			// Parse the JSON into a Resolve.
 			res2, err := DecodeJSON(stdout)
 			if err != nil {